pyproject.toml
[tool.poetry]
name = "blog-restapi-service"
version = "0.1.0"
description = "API service for the blog database"
authors = ["tobias2910 <33369379+tobias2910@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
alembic = "^1.8.1"
asyncpg = "^0.26.0"
bcrypt = "^3.2.2"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python = "^3.9"
psycopg2-binary = "^2.9.3"
pydantic = {extras = ["email"], version = "^1.9.1"}
python-dotenv = "^0.20.0"
python-jose = "^3.3.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.9.5"}
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.8.0"
Faker = "^14.1.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.8.23"
flake8-comprehensions = "^3.10.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.3.0"
flake8-string-format = "^0.3.0"
httpx = "^0.23.0"
isort = "^5.10.1"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-gen-files = "^0.4.0"
mkdocs-material = "^8.4.2"
mkdocs-literate-nav = "^0.4.1"
mkdocs-section-index = "^0.3.4"
mypy = "^0.971"
requests = "^2.28.1"
pep8-naming = "^0.13.2"
pre-commit = "^2.20.0"
pymdown-extensions = "^9.5"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.19.0"
tryceratops = "^1.1.0"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.pyright]
reportGeneralTypeIssues = false
reportUnknownMemberType = false
reportUnknownVariableType = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [
  "."
]

[tool.autoflake]
check = true

[tool.mypy]
exclude = ["alembic/", "tests/"]
disallow_any_unimported = false
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
show_error_codes = true
